Dataset blocks clean up - block templates + duplicated dataset validation (#346)
**NOTE: There is a typo in the branch name - it should be CMS-721** ‼️🚨⚠️

* Removed dead code - unused HTML templates for the DatasetStoryBlock

* Refactored core/utils.py - split into task-specific modules

* Added custom validation for the URL in the DatasetStoryBlock - to ensure consistency with TimeSeriesPageLinkBlock (including checking for duplicates based on the entire url not just the path)
