+++ cms/datasets/blocks.py
@@ -1,5 +1,4 @@
 from collections import defaultdict
-from urllib.parse import urlparse
 
 from django.core.exceptions import ValidationError
 from wagtail.blocks import (
@@ -8,10 +7,13 @@ from wagtail.blocks import (
     StreamBlockValidationError,
     StreamValue,
     StructBlock,
+    StructBlockValidationError,
+    StructValue,
     TextBlock,
     URLBlock,
 )
 
+from cms.core.url_utils import normalise_url, validate_ons_url_struct_block
 from cms.datasets.views import dataset_chooser_viewset
 
 DatasetChooserBlock = dataset_chooser_viewset.get_block_class(
@@ -26,39 +28,40 @@ class ManualDatasetBlock(StructBlock):
 
     class Meta:
         icon = "link"
-        template = "templates/components/streamfield/dataset_link_block.html"
+
+    def clean(self, value: "StructValue") -> "StructValue":
+        errors = validate_ons_url_struct_block(value, self.child_blocks)
+
+        if errors:
+            raise StructBlockValidationError(errors)
+
+        return super().clean(value)
 
 
 class DatasetStoryBlock(StreamBlock):
-    dataset_lookup = DatasetChooserBlock(
-        label="Lookup Dataset", template="templates/components/streamfield/dataset_link_block.html"
-    )
+    dataset_lookup = DatasetChooserBlock(label="Lookup Dataset")
     manual_link = ManualDatasetBlock(
         required=False,
         label="Manually Linked Dataset",
     )
 
-    class Meta:
-        template = "templates/components/streamfield/datasets_block.html"
-
     def clean(self, value: StreamValue, ignore_required_constraints: bool = False) -> StreamValue:
         cleaned_value = super().clean(value)
 
-        # Validate there are no duplicate datasets, including between manual and looked up datasets referencing the same
-        # URL path
+        # Validate there are no duplicate datasets,
+        # including between manual and looked up datasets referencing the same URL
 
-        # For each dataset URL path, record the indices of the blocks it appears in
-        url_paths = defaultdict(set)
+        # For each dataset URL, record the indices of the blocks it appears in
+        urls = defaultdict(set)
         for block_index, block in enumerate(cleaned_value):
-            url_path = (
-                block.value.url_path if block.block_type == "dataset_lookup" else urlparse(block.value["url"]).path
-            )
-            url_paths[url_path].add(block_index)
+            url = block.value.website_url if block.block_type == "dataset_lookup" else block.value["url"]
+            url = normalise_url(url)
+            urls[url].add(block_index)
 
         block_errors = {}
-        for block_indices in url_paths.values():
-            # Add a block error for any index which contains a duplicate a URL path, so that the validation error
-            # messages appear on the actual duplicate entries
+        for block_indices in urls.values():
+            # Add a block error for any index which contains a duplicate URL,
+            # so that the validation error messages appear on the actual duplicate entries
             if len(block_indices) > 1:
                 for index in block_indices:
                     block_errors[index] = ValidationError("Duplicate datasets are not allowed")
