+++ cms/datasets/tests/test_utils.py
@@ -3,9 +3,7 @@ from wagtail.blocks import StreamValue
 
 from cms.datasets.blocks import DatasetStoryBlock
 from cms.datasets.models import Dataset
-from cms.datasets.utils import (
-    format_datasets_as_document_list,
-)
+from cms.datasets.utils import format_datasets_as_document_list
 
 
 class TestUtils(TestCase):
