--- cms/jinja2/templates/components/streamfield/dataset_link_block.html
@@ -1,4 +1,0 @@
-<li class="ons-list__item">
-    <p><a href="{{ value.website_url|default(value.url) }}">{{ value.title }}</a></p>
-    <p>{{ value.description }}</p>
-</li>
