+++ cms/jinja2/templates/pages/release_calendar/release_calendar_page.html
@@ -2,6 +2,7 @@
 
 {% from "components/hero/_macro.njk" import onsHero %}
 {% from "components/table-of-contents/_macro.njk" import onsTableOfContents %}
+{% from "components/document-list/_macro.njk" import onsDocumentList %}
 
 {%- set official_statistics_badge_url = "https://uksa.statisticsauthority.gov.uk/about-the-authority/uk-statistical-system/types-of-official-statistics/" -%}
 
@@ -85,8 +86,17 @@
             {% block release_content %}
                 {% include_block page.content %}
 
-                {% if page.datasets %}
-                    {% include_block page.datasets %}
+                {% if page.dataset_document_list %}
+                    <section id="data" class="spacing">
+                        <h2>{{ _("Data") }}</h2>
+                    {# fmt:off #}
+                        {{ onsDocumentList({
+                                "headingLevel": 3,
+                                "documents": page.dataset_document_list
+                            })
+                        }}
+                    {# fmt:on #}
+                    </section>
                 {% endif %}
 
                 {% with contact_details = page.contact_details%}
